Fix: 뉴스 출력 중 발생 에러 수정
