+++ src/utils/queryOptionsFactory.ts
@@ -3,8 +3,8 @@ import { TBingNewsFilterQueries, TNewsItem, TUserInfo } from '@/types';
 import APIError from './APIError';
 import ERRCODE from '@/constants/errCode';
 import { queryClient } from '@/queries/queryClient';
-import { SCRAP_QUERY_KEYS } from '@/queries/useScrappedNewsList';
-import { getSearchQueryCache } from '@/queries/useBingNewsFetch';
+
+import { fetchNewsListAndConvert, getSearchQueryCache } from '@/queries/useBingNewsFetch';
 import {
   convertToNewsItem,
   isDuplicatedNews,
@@ -34,37 +34,9 @@ export const queryOptionsFactory = {
     bing: {
       list: (filterQueries: TBingNewsFilterQueries) => ({
         queryKey: [QK_PUBLIC, QK_NEWS, QK_BING, 'LIST', { ...filterQueries }],
-        // FIXME: 나중에 api결과 처리하는 부분 어디에 둘지 결정 필요(전부 애매함)
-        queryFn: async ({ pageParam = 1 }) => {
-          const { keyword } = filterQueries;
-          // api 호출
-          const fetchResult = await fetchBingNews(keyword, pageParam);
 
-          // 검색결과 없으면 not found 에러
-          if (fetchResult.value.length === 0) {
-            throw new APIError(ERRCODE.NEWS_FETCH_NOT_FOUND);
-          }
-          // 스크랩 목록
-          const scrappedNewsList = queryClient.getQueryData<TNewsItem[]>(
-            SCRAP_QUERY_KEYS.SCRAP.LIST.queryKey,
-          );
-          // 현재 뉴스데이터
-          const curNewsItems = getSearchQueryCache(filterQueries);
-          // newsItem형식으로 변환
-          const convertedNewsList = fetchResult.value
-            .map((item) => {
-              const isScrapped = setIsScrapped(item.name, scrappedNewsList);
-              const datePublished = parseDateToFormat(item.datePublished);
-              // 중복된 뉴스는 제거(중복된 뉴스 들어오는 경우 있음)
-              const isDuplicated = isDuplicatedNews(item.name, curNewsItems);
-              if (!isDuplicated) {
-                return convertToNewsItem(item, datePublished, keyword, isScrapped);
-              } else {
-                return undefined;
-              }
-            })
-            .filter(Boolean);
-          return convertedNewsList;
+        queryFn: async ({ pageParam = 1 }) => {
+          return await fetchNewsListAndConvert(filterQueries, pageParam);
         },
       }),
     },
